Events, users, itineraries loaded from DB
Events that are associated with the itinerary (i.e.: have the itinerary name as its location) show up on that itinerary screen.

Users appear in the itinerary. Eventually, this should be ONLY users associated with itinerary.

If you uncomment HomeScreen and comment ItineraryScreen in Page(), you can also see itineraries being loaded in (currently all itineraries in database)